contractHunter/requirements.txt
termcolor==2.3.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.24.0
web3>=6.0.0